% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inoculum_species.R
\docType{data}
\name{inoculum_species}
\alias{inoculum_species}
\title{inoculum_species: Inoculum Species}
\format{
A tibble with 44 rows and 2 variables
\describe{
\item{id_inoc}{Identifier for the fungal inoculum.}
\item{species}{Fungal species used as inoculum.}
}
}
\usage{
inoculum_species
}
\description{
Fungal inoculum species used in the experiment.
}
\keyword{datasets}
